frontend/course: untangling "common"
