+++ src/packages/frontend/course/common/flex-panel.tsx
@@ -3,15 +3,12 @@
  *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
  */
 
-import { Rendered, Component } from "../../app-framework";
-
-export class FlexPanel extends Component<{ header: any }> {
-  render(): Rendered {
-    return (
-      <div className={"panel panel-default smc-vfill"}>
-        <div className="panel-heading">{this.props.header}</div>
-        <div className="panel-body smc-vfill">{this.props.children}</div>
-      </div>
-    );
-  }
-}
+export const FlexPanel: React.FC<{ header: any }> = (props) => {
+  const { header, children } = props;
+  return (
+    <div className={"panel panel-default smc-vfill"}>
+      <div className="panel-heading">{header}</div>
+      <div className="panel-body smc-vfill">{children}</div>
+    </div>
+  );
+};
